requirements.txt
django
djangorestframework
pygments
httpie
django-extensions
